StateQueue Refactor
StateQueue (in serial mode) now executes the first state in the list that is has not exited. This way we don't need to keep track of indices, and the whole thing behaves more intuitively.
We can easily add states to the queue without breaking the current state index.
Also, states no longer need to be manually reset to "READY". We just do that when the statequeue itself Enters, which already happens when we want it to.

Added a `loop` property to states so that they run their own `enter` method upon exiting, ensuring they loop themselves correctly.