Massively refactored
Added RelayNodes and RelayMessages.
A RelayMessage is a message with a target node path, a message identifier, and a dict of args.
RelayNodes are an abstract class capable of handling and propagating RelayMessages to their correct destination(s).
States are now based on the RelayNode class.
This Relay system allows us to pass any message down a hierarchy of States. At the moment, only signal-connection messages are handled.
Signals are now much more flexible. We can specify whether we want the connected callable to have the signal's arguments stripped, and we can now add `state: State` to the end of a connected callable and StateCraft will automatically attach it along with the signal's callables, if you didn't ask it to strip them.